S11P11B209-67 Feat: Pay 계좌 등록 구현

+++ backend/src/main/java/com/common/togather/db/repository/PayAccountRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface PayAccountRepository extends JpaRepository<PayAccount, Long> {
+public interface PayAccountRepository extends JpaRepository<PayAccount, Integer> {
     // memberId로 Pay 계좌 찾기
     Optional<PayAccount> findByMemberId(int memberId);
 }
